+++ mobile/test/mobile_compound_models_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:canlifal_social/services/models/mobile_compound_models.dart';
+import 'package:canlifal_social/features/home/data/models/mobile_compound_models.dart';
 
 void main() {
   group('MobileHomeBundle', () {